+++ revert/server_role.sql
@@ -2,6 +2,13 @@
 
 BEGIN;
 
-drop role postgraphql_server;
+do $$
+begin
+    perform true from pg_roles where rolname='postgraphql_server';
+    if found then
+       drop role postgraphql_server;
+    end if;
+end;
+$$;
 
 COMMIT;
